Create TranslationView
