+++ DailyWidget/DailyWidget.swift
@@ -9,36 +9,6 @@ import WidgetKit
 import SwiftUI
 import DailyLibrary
 
-struct TranslationView : View {
-    var entry: TranslationTimelineProvider.Entry
-
-    var body: some View {
-        VStack {
-            HStack {
-                Text(entry.translation.to.text)
-                    .font(.largeTitle)
-                    .fontWeight(.bold)
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.2)
-                Spacer()
-            }
-            HStack {
-                Text(entry.translation.from.text)
-                    .font(.system(.body))
-                Spacer()
-            }
-
-            Spacer()
-
-            HStack {
-                Text(String(entry.translation.to.language.flagEmoji))
-                    .font(.title)
-                Spacer()
-            }
-        }
-    }
-}
-
 struct TranslationWidget: Widget {
     let kind: String = "DailyWidget"
